[WIP] Make profile controller configurable
Make the authoritative controller for seccomp profiles - either
an annotated ConfigMap or the new SeccompProfile CRD - configurable via
an environment variable in the operator deployment. Instead of watching
for both types of resources, the controller now only respects one or the
other, defaulting to ConfigMap since SeccompProfile is in an alpha
state, but also watches for and emits a warning if it sees a resource of
the other type.