fixup: Add SeccompProfiles for default profiles
Adds new SeccompProfile resources as a separate deployment manifest to
replace the ConfigMap profiles when SeccompProfile is the authoritative
profile controller. Also modifies the controller and unit tests to
always use a ".json" extension, since otherwise the version number in
the nginx profile would cause the profile config file not to have an
extension.